feat(cla): handle token 'None' in 'Cla.from_dictionary'
Accept 'None' as a key in the input dictionary, with the behavior
that all associated types will be added to 'Cla.typ', but 'None'
will not appear as a token, and no new validities will associate.
This behavior allows types with no tokens to be added to 'Cla'
through 'from_dictionary', so that the instantiation does not have
to rely on extra arguments.