refactor(cla): make tabular format the point-of-truth for 'Cla'.
Always instantiate 'Cla' from a dictionary, and remove now-redundant
usages of 'Cla.from_dictionary'. Convert all tuple-list invocations
to the new dictionary format. Make 'Cla.table' the source of truth
for checking validity, and remove all methods for mutating the
classificaiton or its underlying table, so that the representation
can be optimized for read operations.