feat(invariant): make 'invariant' an equivalence relation.
Implement invariant on a classification by inheriting 'Invariant'
from 'EqRelation'. Add method 'Set.is_partition' and refactor
validation in 'EqRelation' to use it. Add method 'Cla.agree_all'
so that is easier to check whether two tokens agree on a given
set of types, and use to build the partition associated to an
invariant.