fix(infomorphism): correct axiom check.
Method 'Infomorphism.satisfies_info_axioms' incorrectly indicated
violation of the fundamental infomorphism axiom, probably due to
confusion about the domain, codomain of the two functions when
it was originally written. Correct this problem, and simplify
the implementation of the check, so that it is easier to
read the original definition from the source code.