feat(cla): use index in the sum.
Use each classification's 'index' member to index the types
in a sum classification, so that sum types can be uniquely
identified with their original classification.
Update tests on the sum to use explicit indexes when needed.
Extend equality comparison on 'Index' to allow comparison
with integers, and treat values equal when 'Index.int'
is equal to an integer operand.