feat(value): introduce equable wrapper class.
Add class 'Value', containing an index and a value,
so that arbitrary values can be associated to a
classification index. Make values compare equal only when
both 'name' and 'index' members are equal.
Change equality comparison for 'Index' so that comparison
is made on integer values, because 'uuid' does not directly
implement equality comparison.