refactor(cla): do not index sum tokens.
Remove indexing of tokens when calculating the sum classification,
introduced in:

b97985dcd6010e08a197c6d17d79f664719dc257

because it produces a structure with a large amount of
redundant information, and significantly complicates handling
of the results. Stub out the calculation of colimits, which was
using this change, while fix in progress.